Fix generator beeing too complex and regression in CodeAnalysisPluginTest
